requirements.txt
python-dotenv
pandas
requests
pymongo
openai==0.28.0
Flask
flask-cors
gunicorn
apscheduler
pytesseract==0.3.8
pdfplumber==0.5.28

